Add REST API importer for organizations (#6)
* Add REST API importer for organizations

The importer is designed around with the idea that it should work for diferent data sources. Yet still the importer comes with some default configurations based on the general REST API interface form djangorestframwork. The default configuration assumes the object data are stored in an array under a `results` key, and next page data can be fetch from the link specified in `next` key.

The Organization model has a self reference field, i.e. parent, which must be saved before current organization object is saved. Thus the REST API must provide a way to get needed data for creating the parent organization, either by providing the object data or a url link from where object data can be fetched.

Refs: #234

* Caching used organization classes and data sources

* Allow renaming data source names

* Update README.md for organization data import

* Rename the command for importing organization

The original name `rest_import` is misleading when the app is used in a project.